Move grpc header to `src/endpoints.h`
