+++ src/apps/external_executor/external_executor.cpp
@@ -6,8 +6,8 @@
 #include "ccf/http_consts.h"
 #include "ccf/json_handler.h"
 #include "ccf/kv/map.h"
+#include "endpoints/grpc.h"
 #include "executor_code_id.h"
-#include "grpc.h"
 #include "kv.pb.h"
 
 #define FMT_HEADER_ONLY
